Module atomic_linked_list::atomic_linked_list

A basic atomic linked list. A generic, singly-linked list data structure that is lock-free and uses AtomicPtrs to ensure safety in the face of multithreaded access. Each node remains in the same spot in memory once it is allocated, and will not be reallocated, which allows an external thread to maintain a reference to it safely (but really, only a Weak reference is safe to maintain, to catch possible Node deletion).

New elements are inserted at the head of the list, and then the head’s next pointer is set up to the point to the node that was previously the head. Thus, the head always points to the most recently added node.

Structs

AtomicLinkedList

AtomicLinkedListIter

AtomicLinkedListIterMut