Struct atomic_linked_list::atomic_linked_list::AtomicLinkedList

pub struct AtomicLinkedList<T> { /* fields omitted */ }

Implementations

impl<T> AtomicLinkedList<T>

pub fn new() -> AtomicLinkedList<T>

create a new empty AtomicLinkedList.

pub fn push_front(&self, data: T) -> Result<(), T>

add a new element to the front of the list.

pub fn push_front_timeout(&self, data: T, max_attempts: u64) -> Result<(), T>

add a new element to the front of the list, but will abort if it fails to do so atomically after the given number of attempts.

pub fn iter(&self) -> AtomicLinkedListIter<'_, T>

returns a forward iterator through this linked list.

pub fn iter_mut(&self) -> AtomicLinkedListIterMut<'_, T>

returns a forward iterator through this linked list, allowing mutation of inner elements.